requirements.txt
flask==3.0.2
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
google-generativeai==0.7.1
gunicorn==20.1.0
flask-socketio==5.3.6
Pillow==9.5.0
pymongo==4.8.0
flask-pymongo==2.3.0
pandas==1.5.3
numpy==1.22.0